+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.0.6] - 2026-07-02
+
+### Added
+- `{{seq_cum_energy}}` template helper: monotonic cumulative energy (kWh / kVArh) with variable daily increments
+- `active_energy_kwh` and `reactive_energy_kvarh` fields in [`energy-triangle.yaml`](examples/energy-triangle.yaml)
+
+---
+
 ## [0.0.5] - 2026-07-02
 
 ### Added
@@ -40,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docker image (ghcr.io/iotmertech/mer) and docker-compose
 - GitHub Actions: CI and release workflow (binaries + Docker)
 
-[Unreleased]: https://github.com/iotmertech/iot-data-generator/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/iotmertech/iot-data-generator/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/iotmertech/iot-data-generator/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/iotmertech/iot-data-generator/releases/tag/v0.0.5
 [0.0.1]: https://github.com/iotmertech/iot-data-generator/releases/tag/v0.0.1